Add ruby submodule
To use the ruby class from puppetlabs you have to have that class
available locally, in this case under modules/ruby/manifests/init.pp
(which is the autoloaded location for the ruby class).

I added the puppetlabs-ruby repo as a submdoule in the module_path.

This means you must also specify a module_path in your vagrant file.

The final piece is since you need ruby-dev (the ruby class) to install
anything from the 'gem' provider, I made that dependency explicit in
default.pp.